# Use Object.defineProperty rather than Object.defineProperties because it is faster