--- a/packages/realm/src/Dictionary.ts
+++ b/packages/realm/src/Dictionary.ts
@@ -142,25 +142,23 @@ export class Dictionary<T = unknown> extends Collection<string, T, [string, T], 
 
     const proxied = new Proxy(this, PROXY_HANDLER) as Dictionary<T>;
 
-    Object.defineProperties(this, {
-      [REALM]: {
-        enumerable: false,
-        configurable: false,
-        writable: false,
-        value: realm,
-      },
-      [INTERNAL]: {
-        enumerable: false,
-        configurable: false,
-        writable: false,
-        value: internal,
-      },
-      [HELPERS]: {
-        enumerable: false,
-        configurable: false,
-        writable: false,
-        value: helpers,
-      },
+    Object.defineProperty(this, REALM, {
+      enumerable: false,
+      configurable: false,
+      writable: false,
+      value: realm,
+    });
+    Object.defineProperty(this, INTERNAL, {
+      enumerable: false,
+      configurable: false,
+      writable: false,
+      value: internal,
+    });
+    Object.defineProperty(this, HELPERS, {
+      enumerable: false,
+      configurable: false,
+      writable: false,
+      value: helpers,
     });
 
     return proxied;