--- a/packages/realm/src/Dictionary.ts
+++ b/packages/realm/src/Dictionary.ts
@@ -168,18 +168,18 @@ export class Dictionary<T = unknown> extends Collection<string, T, [string, T], 
    * The representation in the binding.
    * @internal
    */
-  private [REALM]!: Realm;
+  private declare [REALM]: Realm;
 
   /**
    * The representation in the binding.
    * @internal
    */
-  private [INTERNAL]!: binding.Dictionary;
+  private declare [INTERNAL]: binding.Dictionary;
 
   /**
    * @internal
    */
-  private [HELPERS]!: TypeHelpers;
+  private declare [HELPERS]: TypeHelpers;
 
   /** @ts-expect-error We're exposing methods in the end-users namespace of keys */
   [key: string]: T;