Yes, this is kinda stupid, but the results are fairly significant. And, IMO at least, the code is as clear or clearer like this.

FYI, I'm looking into how to get our list perf back up to v11 levels (or better), and this was part of that work. Once I discovered the impact of this, I figured I might as well globally replace all of `Object.defineProperties` and bank that improvement, before doing additional List/OrderedCollection-specific optimization. I'll pick those back up tomorrow.

Edit: I decided to sneak a second related commit in once I remembered that the combination of `prop!: Type` and `Object.declareProperty()` was bad news. The first commit made it more obvious that we were doing that.

Before:
```
    reading property of type 'bool?[]'
      ✓ reads bool?[] (81,881 ops/sec, ±22%)
    reading property of type 'bool?<>'
      ✓ reads bool?<> (108,726 ops/sec, ±14.99%)
    reading property of type 'bool?{}'
      ✓ reads bool?{} (211,068 ops/sec, ±15.59%)
```

After first commit (using `Object.declareProperty`):
```
    reading property of type 'bool?[]'
      ✓ reads bool?[] (90,663 ops/sec, ±29.44%)
    reading property of type 'bool?<>'
      ✓ reads bool?<> (117,569 ops/sec, ±24.64%)
    reading property of type 'bool?{}'
      ✓ reads bool?{} (255,618 ops/sec, ±24.08%)
```

After second commit (using `declare property`):
```
    reading property of type 'bool?[]'
      ✓ reads bool?[] (142,468 ops/sec, ±42.28%)
    reading property of type 'bool?<>'
      ✓ reads bool?<> (185,514 ops/sec, ±28.95%)
    reading property of type 'bool?{}'
      ✓ reads bool?{} (316,215 ops/sec, ±27.6%)
```

## What, How & Why?
<!-- Describe the changes and give some hints to guide your reviewers if possible. -->
<!-- E.g. reference to other repos: This closes realm/realm-sync#??? -->
<!-- Please read CONTRIBUTING.md -->

This closes # ??? <!-- link to an existing issue -->

## ☑️ ToDos
<!-- Add your own todos here -->
* [ ] 📝 Changelog entry
* [ ] 📝 `Compatibility` label is updated or copied from previous entry
* [ ] 📝 Update `COMPATIBILITY.md`
* [ ] 🚦 Tests
* [ ] 🔀 Executed flexible sync tests locally if modifying flexible sync
* [ ] 📦 Updated internal package version in consuming `package.json`s (if updating internal packages)
* [ ] 📱 Check the React Native/other sample apps work if necessary
* [ ] 📝 Public documentation PR created or is not necessary
* [ ] 💥 `Breaking` label has been applied or is not necessary

*If this PR adds or changes public API's:*
* [ ] typescript definitions file is updated
* [ ] jsdoc files updated