@RedBeard0531 2023-05-23
> The numbers are clear, and it makes sense to let them show the path.
> 
> Interestingly, [`Object.defineProperty()` is reported to be slow](https://bugs.chromium.org/p/v8/issues/detail?id=8447#c12) for V8. It doesn't contradict this PR and it hints that `Object.defineProperties()` is much slower than it should be.

I think that bug was before https://v8.dev/blog/faster-class-features. That post was what tipped me off that there might be some new optimizations available for `Object.defineProperties()`, and I decided to check if they weren't all available to `defineProperties()`.

> Do we have numbers for Hermes and JSC?

Not yet. I can get hermes numbers today. Would JSC numbers affect our decision at all?