@RedBeard0531 2023-05-23
> Converting to draft to fix test failures. Hold off on reviewing for now.

Ok, fixed. The issue was that we were using `this.realm` in the constructor prior to setting the property, and I changed it to just `realm` (the argument). I also changed `this.results` to `results` in the `addListener` callback while I was at it, not because it was necessary (since that expression doesn't evaluate until after init completes), but because A) it is easier to just have the rule be to not use `this.blah` in that ctor, and B) there is no reason to use indirection through `this` there, since both `this` and `results` are eagerly bound in the `=>` lambda, so it couldn't possibly evaluate to anything else.