@RedBeard0531 2023-05-23
I was slightly concerned that the issue with `defineProperties` was caused by hitting some perverse case when redefining properties, potentially in a different order, so I tried going back to using `defineProperties`, but with `declare prop`. The results show that both changes make sense, both separately and together (at least on node 18).

```
    reading property of type 'bool?[]'
      ✓ reads bool?[] (116,515 ops/sec, ±34.78%)
    reading property of type 'bool?<>'
      ✓ reads bool?<> (158,146 ops/sec, ±22.99%)
    reading property of type 'bool?{}'
      ✓ reads bool?{} (260,956 ops/sec, ±22.42%)
```

The best results are still with `Object.defineProperty` + `declare prop`. I'll try testing with hermes soon.