@RedBeard0531 2023-05-23
Ok, after much struggle, I now have the hermes numbers, and they are quite _interesting_, but don't suggest any changes to this PR.

Summary:
* Hermes doesn't seem to care much about `declare prop` vs `prop!`. Perf difference on that axis is within the run-to-run noise range. It is possible that it literally makes *no* difference since something in the metro/hermes chain just completely ignores it (my unconfirmed theory)
* Hermes might slightly prefer `declareProperty` for normal properties, but the difference is small and it is hard to be sure it isn't noise.
* Hermes *really* hates `declareProperties` for `Symbol` properties. They are used by Dictionary, and I think that is why there is a huge improvement in this PR for the `bool?{}` case.

`prop!` + `defineProperties` (baseline):
```
    reading property of type 'bool?[]'
      ✓ reads bool?[] (39,783 ops/sec, ±18.25%)
    reading property of type 'bool?<>'
      ✓ reads bool?<> (45,996 ops/sec, ±16.59%)
    reading property of type 'bool?{}'
      ✓ reads bool?{} (51,490 ops/sec, ±16.03%)
```


`prop!` + `defineProperty` (patch 1):
```
    reading property of type 'bool?[]'
      ✓ reads bool?[] (41,343 ops/sec, ±16.7%)
    reading property of type 'bool?<>'
      ✓ reads bool?<> (48,760 ops/sec, ±16.66%)
    reading property of type 'bool?{}'
      ✓ reads bool?{} (82,426 ops/sec, ±19.77%)
```

`declare prop` + `defineProperties`:
```
    reading property of type 'bool?[]'
      ✓ reads bool?[] (40,020 ops/sec, ±14.8%)
    reading property of type 'bool?<>'
      ✓ reads bool?<> (47,643 ops/sec, ±14.1%)
    reading property of type 'bool?{}'
      ✓ reads bool?{} (53,016 ops/sec, ±15.75%)
```

`declare prop` + `defineProperty` (full PR):
```
    reading property of type 'bool?[]'
      ✓ reads bool?[] (41,139 ops/sec, ±20.84%)
    reading property of type 'bool?<>'
      ✓ reads bool?<> (48,287 ops/sec, ±17.62%)
    reading property of type 'bool?{}'
      ✓ reads bool?{} (82,485 ops/sec, ±22.28%)
```


I would really rather not collect all of these numbers for JSC. It would take a long time, and I suspect that we wouldn't change any decisions based on them.